--- src/component/Layout/Layout.tsx
@@ -1,9 +1,0 @@
-import { Header } from './Header';
-
-export function Layout() {
-  return (
-    <>
-      <Header />
-    </>
-  );
-}
